# frontend/public/openapi.json
{
  "components": {
    "schemas": {
...
      "AttestationInput": {
        "properties": {
          "attestationObject": {
            "description": "webauthn: base64url attestationObject from navigator.credentials.create (fmt packed or none)",
            "type": "string"
          },
          "chain": {
            "description": "x509: PEM certificates, the device's own first; must verify to a root in DEVICE_ROOTS_DIR",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "clientDataJSON": {
            "description": "webauthn: base64url clientDataJSON; type webauthn.create, challenge sha256(\"CARGOFLOW-V1-REGISTER\\n\" + shipment id)",
            "type": "string"
          },
          "format": {
            "enum": [
              "x509",
              "webauthn"
            ],
            "type": "string"
          }
        },
        "required": [
          "format"
        ],
        "type": "object"
      },
...
      "Cover": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "financier": {
            "type": "string"
          },
          "financierPayout": {
            "type": "string"
          },
          "insurer": {
            "type": "string"
          },
          "insurerReturn": {
            "type": "string"
          },
          "parametric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Parametric"
              },
              {
                "type": "null"
              }
            ]
          },
          "premium": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "financier",
          "financierPayout",
          "insurer",
          "insurerReturn",
          "premium",
          "status"
        ],
        "type": "object"
      },
...
      "EpochSummary": {
        "properties": {
          "commitTx": {
            "type": "string"
          },
          "compliant": {
            "type": "boolean"
          },
          "conflictBps": {
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "decisionAction": {
            "type": "string"
          },
          "decisionPass": {
            "type": "boolean"
          },
          "endTime": {
            "type": "integer"
          },
          "epochId": {
            "type": "string"
          },
          "heldDistanceM": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "latE6": {
            "type": "integer"
          },
          "lonE6": {
            "type": "integer"
          },
          "maxHumidityX100": {
            "type": "integer"
          },
          "maxShockX100": {
            "type": "integer"
          },
          "milestoneIndex": {
            "type": "integer"
          },
          "penalties": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object"
          },
          "proofVerified": {
            "type": "boolean"
          },
          "readingCount": {
            "type": "integer"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "riskBps": {
            "type": "integer"
          },
          "root": {
            "type": "string"
          },
          "score": {
            "type": "integer"
          },
          "sequence": {
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/EpochSource"
            },
            "type": "array"
          },
          "sourcesTx": {
            "type": "string"
          },
          "startTime": {
            "type": "integer"
          }
        },
        "required": [
          "compliant",
          "conflictBps",
          "createdAt",
          "decisionAction",
          "decisionPass",
          "endTime",
          "epochId",
          "heldDistanceM",
          "latE6",
          "lonE6",
          "maxHumidityX100",
          "maxShockX100",
          "milestoneIndex",
          "penalties",
          "proofVerified",
          "readingCount",
          "reasons",
          "riskBps",
          "root",
          "score",
          "sequence",
          "sources",
          "startTime"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "stable machine-readable code, e.g. invalid_request, unauthorized, replayed, chain_rejected",
                "type": "string"
              },
              "message": {
                "description": "human-readable explanation; never contains internal detail",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FacilityView": {
        "properties": {
          "buyer": {
            "type": "string"
          },
          "closed": {
            "type": "boolean"
          },
          "committed": {
            "type": "string"
          },
          "drawn": {
            "type": "string"
          },
          "exporter": {
            "type": "string"
          },
          "feeBps": {
            "type": "integer"
          },
          "financier": {
            "type": "string"
          },
          "funded": {
            "type": "boolean"
          },
          "milestoneCount": {
            "type": "integer"
          },
          "nextMilestone": {
            "type": "integer"
          },
          "pauseCount": {
            "type": "integer"
          },
          "pauseReason": {
            "type": "string"
          },
          "pausedAt": {
            "type": "integer"
          },
          "remaining": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "vaultPaused": {
            "type": "boolean"
          }
        },
        "required": [
          "buyer",
          "closed",
          "committed",
          "drawn",
          "exporter",
          "feeBps",
          "financier",
          "funded",
          "milestoneCount",
          "nextMilestone",
          "pauseCount",
          "pausedAt",
          "remaining",
          "status",
          "vaultPaused"
        ],
        "type": "object"
      },
...
      "GatewayRequest": {
        "properties": {
          "attestation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AttestationInput"
              },
              {
                "type": "null"
              }
            ],
            "description": "x509 for a secure element's P-256 key; webauthn (required) for a passkey"
          },
          "issuedAt": {
            "type": "integer"
          },
          "keyType": {
            "description": "default ed25519",
            "enum": [
              "ed25519",
              "p256",
              "webauthn"
            ],
            "type": "string"
          },
          "label": {
            "description": "display name, at most 80 characters",
            "type": "string"
          },
          "publicKey": {
            "description": "base64url: a 32-byte Ed25519 key; for p256 and webauthn a SEC1 P-256 point (33 or 65 bytes) or a DER SubjectPublicKeyInfo",
            "type": "string"
          },
          "sensorIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "issuedAt",
          "publicKey",
          "sensorIds",
          "signature"
        ],
        "type": "object"
      },
...
      "Milestone": {
        "properties": {
          "allocatedUsdg": {
            "type": "string"
          },
          "checkpointCommitment": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "evidenceThreshold": {
            "type": "integer"
          },
          "index": {
            "type": "integer"
          },
          "latE6": {
            "type": "integer"
          },
          "lonE6": {
            "type": "integer"
          },
          "placeLabel": {
            "type": "string"
          },
          "radiusM": {
            "type": "integer"
          },
          "releaseTxHash": {
            "type": "string"
          },
          "released": {
            "type": "boolean"
          },
          "releasedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "allocatedUsdg",
          "checkpointCommitment",
          "description",
          "evidenceThreshold",
          "index",
          "latE6",
          "lonE6",
          "placeLabel",
          "radiusM",
          "released"
        ],
        "type": "object"
      },
...
      "Parametric": {
        "properties": {
          "consecutiveFailedEpochs": {
            "type": "integer"
          },
          "epochFloor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "exporterSalvage": {
            "type": "string"
          },
          "salvageToExporter": {
            "type": "string"
          }
        },
        "required": [
          "consecutiveFailedEpochs",
          "epochFloor",
          "exporterSalvage",
          "salvageToExporter"
        ],
        "type": "object"
      },
...
      "Shipment": {
        "properties": {
          "buyer": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "exporter": {
            "type": "string"
          },
          "externalRef": {
            "type": "string"
          },
          "financier": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoiceHash": {
            "type": "string"
          },
          "invoiceValue": {
            "type": "string"
          },
          "placeLabels": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "policyCommitment": {
            "type": "string"
          },
          "route": {
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            },
            "type": "array"
          },
          "routeCommitment": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "buyer",
          "createdAt",
          "exporter",
          "externalRef",
          "id",
          "invoiceHash",
          "invoiceValue",
          "placeLabels",
          "policy",
          "policyCommitment",
          "route",
          "routeCommitment",
          "status",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ShipmentView": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cover"
              },
              {
                "type": "null"
              }
            ]
          },
          "facility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FacilityView"
              },
              {
                "type": "null"
              }
            ]
          },
          "latestEvidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpochSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/Milestone"
            },
            "type": "array"
          },
          "openCoverOffers": {
            "type": "integer"
          },
          "quarantinedReadings": {
            "type": "integer"
          },
          "shipment": {
            "$ref": "#/components/schemas/Shipment"
          },
          "title": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TitleView"
              },
              {
                "type": "null"
              }
            ]
          },
          "usdgDecimals": {
            "type": "integer"
          }
        },
        "required": [
          "cover",
          "facility",
          "latestEvidence",
          "milestones",
          "openCoverOffers",
          "quarantinedReadings",
          "shipment",
          "title",
          "usdgDecimals"
        ],
        "type": "object"
      },
...
      "Source": {
        "properties": {
          "attestation": {
            "additionalProperties": {},
            "description": "what registration verified",
            "type": "object"
          },
          "attested": {
            "description": "an attestation chain verified to a trusted root",
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deviceClass": {
            "enum": [
              "software",
              "passkey",
              "secure_element"
            ],
            "type": "string"
          },
          "deviceTx": {
            "description": "v3: the DeviceRegistry.registerDevice transaction (registration responses only)",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "keyHash": {
            "description": "0x + keccak256(public key bytes); see GET /v1/devices/{keyHash}",
            "type": "string"
          },
          "keyType": {
            "enum": [
              "ed25519",
              "p256",
              "webauthn"
            ],
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "publicKey": {
            "description": "base64url of the stored key: 32-byte Ed25519 or 65-byte uncompressed P-256 point",
            "type": "string"
          },
          "reliabilityBps": {
            "description": "the weight the evidence engine gives this source's sensors",
            "type": "integer"
          },
          "sensorIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "shipmentId": {
            "type": "string"
          }
        },
        "required": [
          "attested",
          "createdAt",
          "deviceClass",
          "id",
          "keyHash",
          "keyType",
          "label",
          "publicKey",
          "reliabilityBps",
          "sensorIds",
          "shipmentId"
        ],
        "type": "object"
      },
...
      "SourceList": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "type": "object"
      },
...
      "TitleView": {
        "properties": {
          "holder": {
            "type": "string"
          },
          "status": {
            "enum": [
              "ISSUED",
              "SURRENDERED",
              "VOID"
            ],
            "type": "string"
          },
          "tokenId": {
            "type": "string"
          }
        },
        "required": [
          "holder",
          "status",
          "tokenId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminKey": {
        "description": "the operator's ADMIN_API_KEY",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "alchemySignature": {
        "description": "Alchemy Notify: hex HMAC-SHA256 of the raw request body with the webhook's signing key (ALCHEMY_WEBHOOK_SIGNING_KEYS)",
        "in": "header",
        "name": "X-Alchemy-Signature",
        "type": "apiKey"
      },
      "sourceSignature": {
        "description": "Signed by a registered evidence source. The signing string is\n\n```\nCARGOFLOW-V1\\n<METHOD>\\n<PATH>\\n<TIMESTAMP>\\n<hex sha256(body)>\n```\n\n`ed25519` sources sign it with Ed25519. `p256` sources sign it with ECDSA P-256 over SHA-256 (X-Signature is the DER or raw 64-byte r||s signature). `webauthn` sources run navigator.credentials.get with challenge = sha256(signing string) (32 raw bytes) and send the assertion: X-Signature = signature, X-WebAuthn-Authenticator-Data = authenticatorData, X-WebAuthn-Client-Data = clientDataJSON (all base64url); the backend checks type `webauthn.get`, the challenge, the registered RP id hash, the user-present flag and the signature over authenticatorData || sha256(clientDataJSON).",
        "in": "header",
        "name": "X-Signature",
        "type": "apiKey"
      },
      "walletSignature": {
        "description": "Not a header: the body carries `issuedAt` and `signature` (hex EIP-191 personal_sign) over the operation's `x-cargoflow-signed-message`. Declared as a scheme so clients can see which operations need a wallet.",
        "in": "header",
        "name": "X-Wallet-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CargoFlow turns signed shipment telemetry into on-chain financing decisions. Public reads need no credentials. Writes are authenticated in one of three ways: the operator's admin key, a party's wallet signature over a fixed message (see each operation's `x-cargoflow-signed-message`), or an evidence source's request signature (Ed25519, P-256 or WebAuthn). Errors are `{\"error\": {\"code\", \"message\"}}` with stable codes. Amounts are USDG base units (6 decimals) as decimal strings.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "CargoFlow API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/shipments/{id}": {
      "get": {
        "operationId": "getShipment",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "Combined store and live chain view of a shipment",
        "tags": [
          "Shipments"
        ]
      }
    },
...
    "/v1/shipments/{id}/sources": {
      "get": {
        "operationId": "listGateways",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceList"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "The shipment's evidence gateways with their device class",
        "tags": [
          "Devices"
        ]
      },
      "post": {
        "description": "Binds a device key to this shipment. `keyType` is `ed25519` (default), `p256` or `webauthn`. A `p256` key with an X.509 `attestation` chain that verifies to a manufacturer root (DEVICE_ROOTS_DIR) is class `secure_element`; a `webauthn` key needs its attestation object (`packed` or `none`) and clientDataJSON (type `webauthn.create`, challenge = sha256(\"CARGOFLOW-V1-REGISTER\\n\" + shipment id)) and is class `passkey`; anything else is `software`. A repeat with the same sensors answers 200; the same key with other sensors is 409; at most 8 per shipment.\n\nSigned by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow evidence source\nshipment: <id>\npublic key: <publicKey as sent>\nsensors: <a,b>\n[key type: <p256|webauthn>  (only for non-ed25519 keys)]\nissued: <t>\n```",
        "operationId": "registerGateway",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "already registered"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Register an evidence gateway (exporter)",
        "tags": [
          "Devices"
        ],
        "x-cargoflow-signed-message": "CargoFlow evidence source\nshipment: <id>\npublic key: <publicKey as sent>\nsensors: <a,b>\n[key type: <p256|webauthn>  (only for non-ed25519 keys)]\nissued: <t>"
      }
    },
...
  },
  "servers": [
    {
      "description": "this backend",
      "url": "/"
    }
  ]
}